requirements.txt
django-classy-tags==0.9.0